\name{Rlazyffi-package}
\alias{Rlazyffi-package}
\alias{Rlazyffi}
\docType{package}
\title{
  Foreign Function Interface for the Lazy
}
\description{
    Foreign Function Interface for the Lazy
}
\author{
Antonio, Fabio Di Narzo
}
\keyword{ package }
\seealso{
\code{\link[Rffi:Rffi-package]{Rffi}}
}
